# api/rest/v1/api.openapi.yaml
openapi: 3.0.2
info:
  title: KICS Results API
  description: TBD
  version: 1.0.0
tags:
  - name: Results
paths:
  '/results/{scan-id}':
    get:
      tags:
        - Results
      summary: get result by scan id
      description: Returns result item for specific scan
      operationId: findResultsByScanID
      parameters:
        - in: path
          name: scan-id
          schema:
            type: string
          required: true
          description: get by Scan ID field
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultItem'
        '500':
          description: Internal Server Error.
  '/scan-summary':
    get:
      tags:
        - Summary
      summary: get summary by scan ids
      description: Returns summary for provided scans
      operationId: findSummariesByScanIDs
      parameters:
        - in: query
          name: scan-ids
          schema:
            type: string
          required: true
          description: scan ids separated by comma
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SummaryItem'
        '400':
          description: Bad request.
        '500':
          description: Internal Server Error.
components:
  schemas:
    SeveritySummaryItem:
      type: object
      properties:
        severity:
          $ref: '#/components/schemas/SeverityEnum'
        counter:
          type: integer
    SummaryItem:
      type: object
      properties:
        scanId:
          type: string
        severityCounters:
          type: array
          items:
            $ref: '#/components/schemas/SeveritySummaryItem'
        totalCounter:
          type: integer
    ResultItem:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        queryID:
          type: string
        queryName:
          type: string
          description: Query name
        severity:
          $ref: '#/components/schemas/SeverityEnum'
        line:
          type: integer
        issueType:
          $ref: '#/components/schemas/IssueTypeEnum'
        searchKey:
          type: string
          description: DOM path in source file
        expectedValue:
          type: string
        actualValue:
          type: string
        value:
          type: string
    IssueTypeEnum:
      type: string
      description: Issue type enum of a result.
      enum:
        - MissingAttribute
        - RedundantAttribute
        - IncorrectValue
    SeverityEnum:
      type: string
      description: Severity enum of a result.
      enum:
        - HIGH
        - MEDIUM
        - LOW
        - INFO
  securitySchemes:
    accessControl:
      type: http
      scheme: bearer
      bearerFormat: JWT